fix(#367): missing mock api get
